refactor(schema): split reference/conversion.rs
split the 2085-line legacy CSL-JSON → InputReference converter
into a conversion/ module with one submodule per reference
category, as prescribed by csl26-rfct item #5:

- conversion/legal.rs (~285): legal-case, statute, regulation,
  treaty, standard, patent, bill, hearing.
- conversion/scholarly.rs (~994): monograph + collection-
  component + edited-book + serial-component + preprint +
  dataset + document + event.
- conversion/media.rs (~145): software, audio-visual.

mod.rs (~767) keeps the shared legacy_* / relation_* / build_*
helpers, the RefContext bundle, the From<Reference> dispatch,
and the existing test suite.

input_reference_from_legacy_edited_book is re-exported from
mod.rs so citum-io's existing import path through
citum_schema::reference::conversion is byte-identical.

no behavior change; 1306/1306 workspace tests pass and clippy
--workspace -D warnings is clean.

closes item #5 on csl26-rfct.
