+++ .beans/csl26-rfct--refactor-priorities-modularization.md
@@ -5,7 +5,7 @@ status: in-progress
 type: milestone
 priority: normal
 created_at: 2026-05-16T14:30:00Z
-updated_at: 2026-05-18T11:03:15Z
+updated_at: 2026-05-18T11:22:45Z
 ---
 
 Analysis of Rust source files (excluding tests) exceeding 800 lines, ranked by refactor priority and grouped by modularization strategy.
@@ -30,9 +30,8 @@ Analysis of Rust source files (excluding tests) exceeding 800 lines, ranked by r
 
 4. [x] **`citum-schema-style/src/locale/mod.rs` (2410 → 1347 lines, -44%)**
    - **Done:** Extracted embedded en-US data to `src/locale/embedded/en_us.rs` (per original target), the raw → Locale conversion to `src/locale/raw_conversion.rs`, and message-ID mappings + MF2 dispatch to `src/locale/message_ids.rs`. 1306/1306 workspace tests pass.
-5. [ ] **`citum-schema-data/src/reference/conversion.rs` (2085 lines)**
-   - **Issues:** Giant legacy CSL-JSON converter.
-   - **Target:** Split by reference category (legal, scholarly, media).
+5. [x] **`citum-schema-data/src/reference/conversion/` (2085 → 767 lines, -63%)**
+   - **Done:** Split into `conversion/{legal,scholarly,media}.rs`; shared helpers, RefContext, and the `From<Reference> for InputReference` dispatch stay in `conversion/mod.rs`. Public re-export of `input_reference_from_legacy_edited_book` keeps the existing `citum_schema::reference::conversion::*` path byte-identical. 1306/1306 workspace tests pass.
 6. [ ] **`citum-migrate/src/upsampler.rs` (2064 lines)**
    - **Issues:** Complex citation position analysis intertwined with upsampling logic.
    - **Target:** Extract `CitationPositionAnalysis` to `src/upsampler/position.rs`.
